+++ src/components/toast-notifications/interfaces.ts
@@ -1,4 +1,12 @@
 //for the function type in any child components that need the function
 type addToastFunction =  (message : string, title: string,  animation : boolean, delay : number, isError : boolean) => void;
 
-export type {addToastFunction}
+interface ToastData {
+    id: number;
+    message: string;
+    title: string;
+    delay?: number;
+    animation?: boolean;
+    isError? : boolean;
+}
+export type {addToastFunction, ToastData}
